refactor: move keygen interfaces
