+++ engine/src/main/java/io/camunda/zeebe/streamplatform/ReplayStateMachine.java
@@ -11,7 +11,6 @@ import io.camunda.zeebe.db.TransactionContext;
 import io.camunda.zeebe.db.ZeebeDbTransaction;
 import io.camunda.zeebe.engine.metrics.ReplayMetrics;
 import io.camunda.zeebe.engine.processing.streamprocessor.RecordValues;
-import io.camunda.zeebe.engine.state.KeyGeneratorControls;
 import io.camunda.zeebe.logstreams.impl.Loggers;
 import io.camunda.zeebe.logstreams.impl.log.LogStreamBatchReaderImpl;
 import io.camunda.zeebe.logstreams.log.LogRecordAwaiter;
@@ -30,6 +29,7 @@ import io.camunda.zeebe.scheduler.retry.RecoverableRetryStrategy;
 import io.camunda.zeebe.scheduler.retry.RetryStrategy;
 import io.camunda.zeebe.streamplatform.api.RecordProcessor;
 import io.camunda.zeebe.streamplatform.api.TypedRecord;
+import io.camunda.zeebe.streamplatform.api.state.KeyGeneratorControls;
 import io.camunda.zeebe.streamplatform.api.state.MutableLastProcessedPositionState;
 import java.util.List;
 import java.util.function.BooleanSupplier;
