+++ engine/src/main/java/io/camunda/zeebe/streamplatform/StreamProcessorContext.java
@@ -9,7 +9,6 @@ package io.camunda.zeebe.streamplatform;
 
 import io.camunda.zeebe.db.TransactionContext;
 import io.camunda.zeebe.engine.processing.streamprocessor.RecordValues;
-import io.camunda.zeebe.engine.state.KeyGeneratorControls;
 import io.camunda.zeebe.logstreams.log.LogStream;
 import io.camunda.zeebe.logstreams.log.LogStreamBatchWriter;
 import io.camunda.zeebe.logstreams.log.LogStreamReader;
@@ -21,6 +20,7 @@ import io.camunda.zeebe.streamplatform.api.InterPartitionCommandSender;
 import io.camunda.zeebe.streamplatform.api.ProcessingScheduleService;
 import io.camunda.zeebe.streamplatform.api.ReadonlyStreamProcessorContext;
 import io.camunda.zeebe.streamplatform.api.TypedRecord;
+import io.camunda.zeebe.streamplatform.api.state.KeyGeneratorControls;
 import io.camunda.zeebe.streamplatform.api.state.MutableLastProcessedPositionState;
 import java.util.function.BooleanSupplier;
