+++ engine/src/main/java/io/camunda/zeebe/streamplatform/api/RecordProcessorContext.java
@@ -10,8 +10,8 @@ package io.camunda.zeebe.streamplatform.api;
 import io.camunda.zeebe.db.TransactionContext;
 import io.camunda.zeebe.db.ZeebeDb;
 import io.camunda.zeebe.engine.state.EventApplier;
-import io.camunda.zeebe.engine.state.KeyGenerator;
 import io.camunda.zeebe.engine.state.mutable.MutableZeebeState;
+import io.camunda.zeebe.streamplatform.api.state.KeyGenerator;
 import java.util.List;
 import java.util.function.Function;
