+++ engine/src/main/java/io/camunda/zeebe/streamplatform/api/state/KeyGeneratorControls.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.engine.state;
+package io.camunda.zeebe.streamplatform.api.state;
 
 /** Allows to manipulate the key generator. Should be used with caution. */
 public interface KeyGeneratorControls extends KeyGenerator {
